requirements.txt
cffi==1.2.1
characteristic==14.3.0
ConcurrentLogHandler>=0.9.1
cryptography==0.9.1
cssselect==0.9.1
enum34==1.0.4
funcsigs==0.4
functools32==3.2.3-2
idna==2.0
ipaddress==1.0.14
jsonschema==2.5.1
kafka-python==0.9.5
kazoo>=2.2.1
lxml==3.4.4
mock>=1.3.0
nose==1.3.7
pbr==1.8.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==0.15.1
python-json-logger>=0.1.2
PyYAML==3.11
py-getch==1.0.1
queuelib==1.2.2
redis>=2.10.3
Scrapy==1.0.5
scutils==1.1
service-identity==14.0.0
six==1.10.0
testfixtures>=4.3.3
tldextract==1.7.1
Twisted==15.4.0
w3lib==1.12.0
wheel==0.26.0
zope.interface==4.1.3
